eina_matrix: replace cosf by cos to gain more accuracy
Summary:
cos function is much much more accurate than cosf.
this patch replaces cosf by cos to gain more accuracy.

Reviewers: cedric, jsuya, vtorri

Subscribers: vtorri, #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D10695

+++ src/lib/eina/eina_matrix.c
@@ -551,8 +551,8 @@ eina_matrix3_rotate(Eina_Matrix3 *m, double rad)
     *       Later we would want someone to look at this and improve accuracy.
     */
 #if 1
-   c = cosf(rad);
-   s = sinf(rad);
+   c = cos(rad);
+   s = sin(rad);
 #else
    /* normalize the angle between -pi,pi */
    rad = fmod(rad + M_PI, 2 * M_PI) - M_PI;
